refactor: update stringifier file name
