+++ src/core/stringifier.ts
